principle name: Polymorphism
Polymorphism: is considered one of the important features of Object-Oriented Programming. Polymorphism allows us to perform a single action in different ways. In other words, polymorphism allows you to define one interface and have multiple implementations. The word “poly” means many and “morphs” means forms, so it means many forms. it helps in achieving flexibility and reusability in the code.